Updated Diffusion Examples to 0.0.0-123464.cd5037f9-SNAPSHOT

+++ apple/Features/TimeSeries/TimeSeriesAppendExample.m
@@ -104,7 +104,7 @@
         }
 
         NSString *const author = eventMetadata.author;
-        NSDate *const timestamp = [NSDate dateWithTimeIntervalSince1970:eventMetadata.timestamp / 1000];
+        NSDate *const timestamp = [NSDate dateWithTimeIntervalSince1970:(double)eventMetadata.timestamp / 1000.0];
         NSDateFormatter *const dateFormatter = [[NSDateFormatter alloc] init];
         [dateFormatter setDateFormat:@"yyyy-MM-dd HH:mm:ss"];
         NSLog(@"Time series topic appended by [%@] at [%@]", author, [dateFormatter stringFromDate:timestamp]);
